Partial refactor of OrgSync, and fetching details on issues.
